+++ maps-sdk/src/main/java/com/mapbox/maps/MapboxMap.kt
@@ -1363,6 +1363,24 @@ class MapboxMap :
   }
 
   /**
+   * Queries the map for rendered raster values at a specific coordinate.
+   *
+   * @param coordinate The position on the screen to query.
+   * @param options The options for configuring the rendered raster value query.
+   * @param callback Callback invoked when the query completes
+   * @return A `cancelable` object that could be used to cancel the pending query.
+   */
+  @MapboxExperimental
+  override fun queryRenderedRasterValues(
+    coordinate: ScreenCoordinate,
+    options: RenderedRasterQueryOptions,
+    callback: QueryRenderedRasterValuesCallback
+  ): Cancelable {
+    checkNativeMap("queryRenderedRasterValues", false)
+    return nativeMap.queryRenderedRasterValues(coordinate, options, callback)
+  }
+
+  /**
    * In some cases querying source / render features is expected to be a blocking operation
    * e.g. performing this action on map click. In this case in order to avoid deadlock on main
    * thread querying could be performed on render thread and in that case querying result will be also